+++ app/jobs/warm_entry_cache_job.rb
@@ -5,10 +5,17 @@ class WarmEntryCacheJob < ApplicationJob
   sidekiq_options retry: 5
 
   def perform
-    entries = GetAllContentfulEntries.new.call
+    entries = BuildJourneyOrder.new(
+      entries: GetAllContentfulEntries.new.call,
+      starting_entry_id: ENV["CONTENTFUL_PLANNING_START_ENTRY_ID"]
+    ).call
+
     entries.each do |entry|
       store_in_cache(cache: cache, key: "contentful:entry:#{entry.id}", entry: entry)
     end
+  rescue BuildJourneyOrder::RepeatEntryDetected,
+    BuildJourneyOrder::TooManyChainedEntriesDetected
+    cache.extend_ttl_on_all_entries
   end
 
   private